+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this project will be documented in this file. This project uses [Semantic Versioning](https://semver.org/).
 
+## v3.0.3
+* Update node-version: [18.x, 20.x, 22.x], remove 16.x which is no longer supported by homebridge.
+* Reformated `getStationObservation()` and `getStationCurrentObservation()` in `tempestApi.ts`.
+* Addresses `observation_data is undefined, skipping update` error in `platform.ts` polling loop.
+
 ## v3.0.2
 * Update node-version: [16.x, 18.x, 20.x], remove 14.x which is no longer supported by homebridge.
 * Update `devDependencies` and `dependencies` to latest versions. Update/lock `axios` to version `1.5.1`.